+++ fpsim/locations/__init__.py
@@ -1,2 +1,3 @@
 from . import kenya
 from . import senegal
+from . import ethiopia